backend/requirements.txt
Django==2.2.5
graphene-django==2.5.0 
django-cors-headers==3.1.0     
graphene==2.1.8
djangorestframework==3.10.3 
Markdown==3.1.1
psycopg2==2.7.3.2
